+++ app/server/endpoint/src/main/java/io/syndesis/server/endpoint/v1/handler/integration/IntegrationSpecificationHandler.java
@@ -15,13 +15,10 @@
  */
 package io.syndesis.server.endpoint.v1.handler.integration;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.Consumes;
@@ -42,12 +39,7 @@ import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.ResponseHeader;
-import io.syndesis.common.model.DataShape;
-import io.syndesis.common.model.Kind;
-import io.syndesis.common.model.ResourceIdentifier;
-import io.syndesis.common.model.integration.Flow;
 import io.syndesis.common.model.integration.Integration;
-import io.syndesis.common.model.integration.Step;
 import io.syndesis.common.model.openapi.OpenApi;
 import io.syndesis.integration.api.IntegrationResourceManager;
 import io.syndesis.server.api.generator.APIGenerator;
@@ -112,7 +104,7 @@ public final class IntegrationSpecificationHandler {
         fetch(@NotNull @PathParam("id") @ApiParam(required = true, example = "integration-id", value = "The ID of the integration") final String id) {
         final Integration integration = integrationHandler.getIntegration(id);
 
-        return specificationFrom(integration)
+        return ApiGeneratorHelper.specificationFrom(resourceManager, integration)
             .map(IntegrationSpecificationHandler::createResponseFrom)
             .orElse(NOT_FOUND);
     }
@@ -130,9 +122,9 @@ public final class IntegrationSpecificationHandler {
 
         final Integration givenIntegration = apiIntegration.getIntegration();
 
-        final Integration updated = updateFlowsAndStartAndEndDataShapes(existing, givenIntegration);
+        final Integration updated = ApiGeneratorHelper.updateFlowsAndStartAndEndDataShapes(existing, givenIntegration);
 
-        final OpenApi existingApiSpecification = specificationFrom(existing).orElse(null);
+        final OpenApi existingApiSpecification = ApiGeneratorHelper.specificationFrom(resourceManager, existing).orElse(null);
 
         if (Objects.equals(existing.getFlows(), updated.getFlows()) && Objects.equals(existingApiSpecification, apiIntegration.getSpec())) {
             // no changes were made to the flows or to the specification
@@ -146,15 +138,6 @@ public final class IntegrationSpecificationHandler {
 
         // perform the regular update
         integrationHandler.update(id, updated);
-    }
-
-    Optional<OpenApi> specificationFrom(final Integration integration) {
-        final Optional<ResourceIdentifier> specification = integration.getResources().stream().filter(r -> r.getKind() == Kind.OpenApi)
-            .max(ResourceIdentifier.LATEST_VERSION);
-
-        return specification
-            .flatMap(s -> s.getId()
-                .flatMap(resourceManager::loadOpenApiDefinition));
     }
 
     static Response createResponseFrom(final OpenApi document) {
@@ -171,64 +154,4 @@ public final class IntegrationSpecificationHandler {
             .build();
     }
 
-    static Integration updateFlowsAndStartAndEndDataShapes(final Integration existing, final Integration given) {
-        // will contain updated flows
-        final List<Flow> updatedFlows = new ArrayList<>(given.getFlows().size());
-
-        for (final Flow givenFlow : given.getFlows()) {
-            final String flowId = givenFlow.getId().get();
-
-            final Optional<Flow> maybeExistingFlow = existing.findFlowById(flowId);
-            if (!maybeExistingFlow.isPresent()) {
-                // this is a flow generated from a new operation or it
-                // has it's operation id changed, either way we only
-                // need to add it, since we don't know what flow we need
-                // to update
-                updatedFlows.add(givenFlow);
-                continue;
-            }
-
-            final List<Step> givenSteps = givenFlow.getSteps();
-            if (givenSteps.size() != 2) {
-                throw new IllegalArgumentException("Expecting to get exactly two steps per flow");
-            }
-
-            // this is a freshly minted flow from the specification
-            // there should be only two steps (start and end) in the
-            // flow
-            final Step givenStart = givenSteps.get(0);
-            final Optional<DataShape> givenStartDataShape = givenStart.outputDataShape();
-
-            // generated flow has only a start and an end, start is at 0
-            // and the end is at 1
-            final Step givenEnd = givenSteps.get(1);
-            final Optional<DataShape> givenEndDataShape = givenEnd.inputDataShape();
-
-            final Flow existingFlow = maybeExistingFlow.get();
-            final List<Step> existingSteps = existingFlow.getSteps();
-
-            // readability
-            final int start = 0;
-            final int end = existingSteps.size() - 1;
-
-            // now we update the data shapes of the start and end steps
-            final Step existingStart = existingSteps.get(start);
-            final Step updatedStart = existingStart.updateOutputDataShape(givenStartDataShape);
-
-            final Step existingEnd = existingSteps.get(end);
-            final Step updatedEnd = existingEnd.updateInputDataShape(givenEndDataShape);
-
-            final List<Step> updatedSteps = new ArrayList<>(existingSteps);
-            updatedSteps.set(start, updatedStart);
-            updatedSteps.set(end, updatedEnd);
-
-            final Flow updatedFlow = existingFlow.builder().name(givenFlow.getName()).description(givenFlow.getDescription()).steps(updatedSteps).build();
-            updatedFlows.add(updatedFlow);
-        }
-
-        return existing.builder().flows(updatedFlows)
-            // we replace all resources counting that the only resource
-            // present is the OpenAPI specification
-            .resources(given.getResources()).build();
-    }
 }
